Add freespace layer
The freespace layer computes the free space above the meshes vertices
and marks vertices as lethal if the robot would not fit through e.g. a
doorway.

Add exporting of cmake targets to mesh_map for depending projects to
use.

LVR2_DEFINITIONS had to be added to the mesh_map target to reflect LVR2 library configuration.