Add ObstacleLayer
This layer maps obstacles represented by PointCloud2 messages onto the
mesh surface using raycasting. The vertices of the obstructed faces are
added to the lethal vertex set and their cost is set to infinity. This
enables replanning around dynamic and unmapped obstacles and dynamic
obstacle avoidance (support depends on the controller plugin).