Remove floating point range from lethal and inscribed cost parameters.
rqt_reconfigure was unable to fetch the parameters when the cost was set
to inf. Since we want to allow setting the cost values to inf and the
range was only there to allow rqt_reconfigure to show a slider (which
did not work with inf) we decided to remove the range.